Move `ProtocConfigurationPlugin` under `plugin-base` to allow extensions
Also move all of the supporting utilities (`gradle` package).
