+++ tools/plugin-base/src/main/java/io/spine/tools/gradle/ProtocConfigurationPlugin.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle.compiler;
+package io.spine.tools.gradle;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.gradle.ExecutableLocator;
@@ -27,13 +27,6 @@ import com.google.protobuf.gradle.ProtobufConfigurator;
 import com.google.protobuf.gradle.ProtobufConfigurator.GenerateProtoTaskCollection;
 import com.google.protobuf.gradle.ProtobufConvention;
 import io.spine.code.fs.java.DefaultJavaProject;
-import io.spine.code.proto.DescriptorReference;
-import io.spine.tools.gradle.Artifact;
-import io.spine.tools.gradle.GradleTask;
-import io.spine.tools.gradle.ProtobufDependencies;
-import io.spine.tools.gradle.SourceScope;
-import io.spine.tools.gradle.SpinePlugin;
-import io.spine.tools.gradle.TaskName;
 import io.spine.tools.groovy.GStrings;
 import org.gradle.api.Action;
 import org.gradle.api.GradleException;
@@ -42,46 +35,35 @@ import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.Dependency;
-import org.gradle.api.plugins.JavaPluginConvention;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Base64;
 import java.util.Collection;
 
-import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.code.fs.java.DefaultJavaProject.at;
 import static io.spine.tools.gradle.ConfigurationName.FETCH;
 import static io.spine.tools.gradle.ProtobufDependencies.protobufCompiler;
-import static io.spine.tools.gradle.TaskName.clean;
-import static io.spine.tools.gradle.TaskName.writeDescriptorReference;
-import static io.spine.tools.gradle.TaskName.writePluginConfiguration;
-import static io.spine.tools.gradle.TaskName.writeTestDescriptorReference;
-import static io.spine.tools.gradle.TaskName.writeTestPluginConfiguration;
-import static io.spine.tools.gradle.compiler.Extension.getMainDescriptorSet;
-import static io.spine.tools.gradle.compiler.Extension.getTestDescriptorSet;
 import static io.spine.tools.groovy.ConsumerClosure.closure;
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static org.gradle.internal.os.OperatingSystem.current;
 
 /**
- * The Gradle plugin which configures Protobuf compilation.
+ * An abstract base for Gradle plugins that configure Protobuf compilation.
  *
- * <p>This plugin requires {@code com.google.protobuf} plugin. If it is not applied, the plugin
- * performs no action.
+ * <p>Any extending plugin requires {@code com.google.protobuf} plugin. If it is not applied,
+ * no action is performed.
  */
-public class ProtocConfigurationPlugin extends SpinePlugin {
+public abstract class ProtocConfigurationPlugin extends SpinePlugin {
 
     private static final String PLUGIN_ARTIFACT_PROPERTY = "Protoc plugin artifact";
 
     private static final String GRPC_GROUP = "io.grpc";
     private static final String GRPC_PLUGIN_NAME = "protoc-gen-grpc-java";
 
-    private static final String SPINE_PLUGIN_NAME = "spine-protoc-plugin";
+    protected static final String SPINE_PLUGIN_NAME = "spine-protoc-plugin";
     private static final String JAR_EXTENSION = "jar";
     private static final String SH_EXTENSION = "sh";
     private static final String BAT_EXTENSION = "bat";
@@ -92,7 +74,8 @@ public class ProtocConfigurationPlugin extends SpinePlugin {
     @Override
     public void apply(Project project) {
         project.getPluginManager()
-               .withPlugin(ProtobufDependencies.gradlePlugin().value(), plugin -> applyTo(project));
+               .withPlugin(ProtobufDependencies.gradlePlugin()
+                                               .value(), plugin -> applyTo(project));
     }
 
     private void applyTo(Project project) {
@@ -201,99 +184,33 @@ public class ProtocConfigurationPlugin extends SpinePlugin {
         boolean tests = isTestsTask(protocTask);
         Project project = protocTask.getProject();
         File descriptor;
-        TaskName writeRefName;
-        if (tests) {
-            descriptor = getTestDescriptorSet(project);
-            writeRefName = writeTestDescriptorReference;
-        } else {
-            descriptor = getMainDescriptorSet(project);
-            writeRefName = writeDescriptorReference;
-        }
+        descriptor = tests
+                     ? getTestDescriptorSet(project)
+                     : getMainDescriptorSet(project);
         GenerateProtoTask.DescriptorSetOptions options = protocTask.getDescriptorSetOptions();
         options.setPath(GStrings.fromPlain(descriptor.getPath()));
         options.setIncludeImports(true);
         options.setIncludeSourceInfo(true);
 
-        JavaPluginConvention javaConvention = project.getConvention()
-                                                     .getPlugin(JavaPluginConvention.class);
-        SourceScope sourceScope = tests ? SourceScope.test : SourceScope.main;
-        Path resourceDirectory = descriptor.toPath()
-                                           .getParent();
-        javaConvention.getSourceSets()
-                      .getByName(sourceScope.name())
-                      .getResources()
-                      .srcDir(resourceDirectory);
-        GradleTask writeRef = newTask(writeRefName, task -> {
-            DescriptorReference reference = DescriptorReference.toOneFile(descriptor);
-            reference.writeTo(resourceDirectory);
-        }).allowNoDependencies()
-          .applyNowTo(project);
-        protocTask.finalizedBy(writeRef.getTask());
+        configureDescriptorSetGeneration(protocTask, descriptor);
     }
 
-    private void configureTaskPlugins(GenerateProtoTask protocTask, Task dependency) {
-        Path spineProtocConfigPath = spineProtocConfigPath(protocTask);
-        Task writeConfig = newWriteSpineProtocConfigTask(protocTask, spineProtocConfigPath);
-        protocTask.dependsOn(dependency, writeConfig);
-        protocTask.getPlugins()
-                  .create(ProtocPlugin.GRPC.name);
-        protocTask.getPlugins()
-                  .create(ProtocPlugin.SPINE.name,
-                          options -> {
-                              options.setOutputSubDir("java");
-                              String option = spineProtocConfigPath.toString();
-                              String encodedOption = base64Encoded(option);
-                              options.option(encodedOption);
-                          });
-    }
+    protected abstract File getMainDescriptorSet(Project project);
 
-    private static String base64Encoded(String value) {
-        Base64.Encoder encoder = Base64.getEncoder();
-        byte[] valueBytes = value.getBytes(UTF_8);
-        String result = encoder.encodeToString(valueBytes);
-        return result;
-    }
+    protected abstract File getTestDescriptorSet(Project project);
 
-    /**
-     * Creates a new {@code writeSpineProtocConfig} task that is expected to run after the
-     * {@code clean} task.
-     */
-    private Task newWriteSpineProtocConfigTask(GenerateProtoTask protocTask, Path configPath) {
-        return newTask(spineProtocConfigWriteTaskName(protocTask), task -> {
-            ProtocPluginConfiguration configuration = ProtocPluginConfiguration
-                    .forProject(protocTask.getProject());
-            configuration.writeTo(configPath);
-        }).allowNoDependencies()
-          .applyNowTo(protocTask.getProject())
-          .getTask()
-          .mustRunAfter(clean.value());
-    }
+    protected abstract void configureDescriptorSetGeneration(GenerateProtoTask task,
+                                                             File descriptor);
 
-    private static TaskName spineProtocConfigWriteTaskName(GenerateProtoTask protoTask) {
-        return isTestsTask(protoTask) ?
-               writeTestPluginConfiguration :
-               writePluginConfiguration;
-    }
+    protected abstract void configureTaskPlugins(GenerateProtoTask protocTask, Task dependency);
 
-    private static Path spineProtocConfigPath(GenerateProtoTask protocTask) {
-        Project project = protocTask.getProject();
-        File buildDir = project.getBuildDir();
-        Path spinePluginTmpDir = Paths.get(buildDir.getAbsolutePath(),
-                                           "tmp",
-                                           SPINE_PLUGIN_NAME);
-        Path protocConfigPath = isTestsTask(protocTask) ?
-                                spinePluginTmpDir.resolve("test-config.pb") :
-                                spinePluginTmpDir.resolve("config.pb");
-        return protocConfigPath;
-    }
-
-    private static boolean isTestsTask(GenerateProtoTask protocTask) {
+    protected static boolean isTestsTask(GenerateProtoTask protocTask) {
         return protocTask.getSourceSet()
                          .getName()
                          .contains(SourceScope.test.name());
     }
 
-    private enum ProtocPlugin {
+    protected enum ProtocPlugin {
 
         GRPC("grpc"),
         SPINE("spineProtoc");
@@ -303,6 +220,10 @@ public class ProtocConfigurationPlugin extends SpinePlugin {
 
         ProtocPlugin(String name) {
             this.name = name;
+        }
+
+        public String getName() {
+            return name;
         }
     }
