+++ tools/plugin-base/src/main/java/io/spine/tools/groovy/GStrings.java
