+++ tools/plugin-base/src/main/java/io/spine/tools/groovy/package-info.java
